Removed and changed significant parts of the library

+++ examples/WIFConverter.c
@@ -26,7 +26,7 @@ int main(int argc, char * argv[]){
 	bool useCompression = CBWIFUseCompression(&wif);
 
 	// Get key
-	uint8_t key[32];
+	int key[32];
 	CBWIFGetPrivateKey(&wif, key);
 	CBDestroyWIF(&wif);
